+++ NEWS.md
@@ -1,6 +1,10 @@
 # Version (development version)
 
- * ...
+## Bug Fixes
+
+ * `availableCores()` would produce a warning `In is.na(constraints) :
+   is.na() applied to non-(list or vector) of type 'NULL'` when
+   running with R (< 4.0.0).
  
 
 # Version 1.33.0 [2022-12-13]